fix(desktop): move language setting to general and dedupe plugin settings
