+++ apps/desktop/src/components/settings/sections/appearance_section.tsx
@@ -13,12 +13,6 @@ function AppearanceSection() {
     { value: "dark", label: t("settings.appearance.theme.dark") },
     { value: "light", label: t("settings.appearance.theme.light") },
   ];
-  const languageOptions = [
-    { value: "system", label: t("settings.appearance.language.system") },
-    { value: "en", label: t("settings.appearance.language.en") },
-    { value: "ko", label: t("settings.appearance.language.ko") },
-    { value: "ja", label: t("settings.appearance.language.ja") },
-  ];
 
   return (
     <SettingsPanel
@@ -26,22 +20,6 @@ function AppearanceSection() {
       description={t("settings.appearance.description")}
       anchor="appearance"
     >
-      <SettingsFieldRow
-        label={t("settings.appearance.language.label")}
-        description={t("settings.appearance.language.description")}
-        control={
-          <div class="w-full max-w-56">
-            <SettingsSelect
-              options={languageOptions}
-              value={settingsState.appearance.language}
-              onChange={(value) =>
-                setAppearanceSetting("language", value as "system" | "en" | "ko" | "ja")
-              }
-              placeholder={t("settings.appearance.language.placeholder")}
-            />
-          </div>
-        }
-      />
       <SettingsFieldRow
         label={t("settings.appearance.theme.label")}
         description={t("settings.appearance.theme.description")}
