+++ apps/desktop/src/components/settings/sections/general_section.tsx
@@ -6,9 +6,10 @@ import {
   SettingsStatusBadge,
   SettingsToolbarAction,
   SettingsCard,
+  SettingsSelect,
 } from "~/components/settings/settings_blocks";
 import { t } from "~/i18n";
-import { settingsState } from "~/stores/settings";
+import { setAppearanceSetting, settingsState } from "~/stores/settings";
 import { clearConfiguredVault, selectVault } from "~/stores/vault";
 
 function VaultFolderControl() {
@@ -78,6 +79,26 @@ function VaultFolderControl() {
   );
 }
 
+function LanguageControl() {
+  const languageOptions = [
+    { value: "system", label: t("settings.appearance.language.system") },
+    { value: "en", label: t("settings.appearance.language.en") },
+    { value: "ko", label: t("settings.appearance.language.ko") },
+    { value: "ja", label: t("settings.appearance.language.ja") },
+  ];
+
+  return (
+    <div class="w-full max-w-56">
+      <SettingsSelect
+        options={languageOptions}
+        value={settingsState.appearance.language}
+        onChange={(value) => setAppearanceSetting("language", value as "system" | "en" | "ko" | "ja")}
+        placeholder={t("settings.appearance.language.placeholder")}
+      />
+    </div>
+  );
+}
+
 function GeneralSection() {
   return (
     <SettingsPanel
@@ -85,6 +106,11 @@ function GeneralSection() {
       description={t("settings.general.description")}
       anchor="general"
     >
+      <SettingsFieldRow
+        label={t("settings.appearance.language.label")}
+        description={t("settings.appearance.language.description")}
+        control={<LanguageControl />}
+      />
       <SettingsFieldRow
         stacked
         label={t("settings.general.vault_folder.label")}
